refactor(form): reorganize components into Form directory
Consolidate all form-related components into a unified Form directory
structure with consistent naming:
- FormAdvancedLink and FormAdvancedLinkTab to Form/FormAdvancedLink/
- FormDonate, FormEmbed, FormSignUp to Form/
- FormControl components to Form/FormControl/

Rename components to follow consistent naming convention:
- AdvancedLinkForm → FormAdvancedLink
- DonateForm → FormDonate
- EmbedForm → FormEmbed
- SignUpForm → FormSignUp
- DigitsInput → FormControlDigits
- RangePicker → FormControlRange
- SecretInput → FormControlSecret
- SelectableDropdown → FormControlSelectableDropdown
