+++ lib/components/Form/FormControl/FormControlRange.vue
